[489904]
Fix for bug 489904

If a DSL file is only partially formatted it is possible, that the
ITextReplacement is not fully contained in an allowed ITextRegion but
they overlap.

Modified method 'TextReplacerContext.isInRequestedRange(...)' to not
only check for containment but also for intersection. To check for
intersection I created a private method
'TextReplacerContext.doIntersect(...)'.

Additionally added a test with the example language and formatter
described in the bug.

Signed-off-by: cwehrheim <cwehrheim@itemis.de>