[bug#489904] Fix partial formatting on single line
If a DSL file is only partially formatted it is possible that the
ITextReplacement is not fully contained in an allowed ITextRegion but
they overlap.

Modified method 'TextReplacerContext.isInRequestedRange(...)' to not
only check for containment but also for intersection. To check for
intersection a private method
'TextReplacerContext.doIntersect(...)' was added.

Added PartialFormattingTest test case and formatter
described in the bug.

Signed-off-by: Christian Wehrheim <christian.wehrheim@itemis.de>
Also-by: Karsten Thoms <karsten.thoms@itemis.de>