Intern grammar elements when nodes are created from the cache
The node model contains nodes derived from AbstractNode.  An
AbstractNode has an array describing its grammar elements.  This array
is often identical for many AbstractNodes.  There can be millions of
nodes in a node model so it makes sense to remove the duplicate arrays.
This GrammarElementsInterner tries to reuse an interned version of
such an array wherever possible.  In particular, now we also intern
grammar elements when they are created during the load of a node model
from the node model cache.

Testing this modification on a largisch project showed that the
execution time of a clean build that accesses the node model cache
reduces by around 10%.

Signed-off-by: Mark Christiaens <mark.christiaens@sigasi.com>