[#1304] Refactor retrieval of the own manifest URL
This changes the way the MANIFEST.MF for org.eclipse.xtext.util is
located. It uses the classloader to locate the class resource of
XtextVersion and resolves the URL relative to the resource. It considers
the case that the class resource can be either on file system or
packaged in an archive, which varies the relative location.

Add a test case for XtextVersion#getCurrent() to cover the regression
that arised by the recent refactoring from Xtend to Java.

Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>