[https://github.com/eclipse/xtext-eclipse/issues/1221] Serializer.
- Improve the serializeReplacement implementation by modifying the
ReplaceRegion length calculation so that it takes not only the
ICompositeNode.getTotalLength() into account, but considers if the new
text contains additional whitespaces and the old text is also followed
by white spaces, than the whitespaces contained by the original document
is also consumed by the quickfix.
- Implement corresponding SerializerReplacementCalculationTest test case
based on the NoJdtTestLanguage.

Signed-off-by: Tamas Miklossy <miklossy@itemis.de>