Replace and transfer look ahead with offset
NodeModelBuilder when adjusting offsets performs it for the whole node
model starting from the root. This is necessary for partial reparse as
the offsets of the following nodes could change.

For parser delegation, that a node corresponding to a terminal rule is
is replaced by new root node (where the node corresponding to the
terminal has been parsed by a different language inside a value
converter), doing the full adjustments leads to a big overhead with
large number of delegations, since the initial offset is know, the
replacement is done directly in the node which needs reparsing and does
not start from the root.

For this use case, we create a new method which keeps the offset of the
old node and iterates over just the tree under from new node.