Build of persisted state to run in parallel with work on dirty state.
Enable a build of the persisted state to run in parallel to work
performed on the dirty state by splitting the index, the resource set,
the content provider, and the request manager into persisted and dirty
components.

Introduces a limited scope dirty build so that validations are performed
for dirty editors and the effect of changes in one dirty editor can be
seen in other dirty editors.